feat: enhance user controller and service with validation and error handling

+++ src/users/users.controller.ts
@@ -6,10 +6,12 @@ import {
   Patch,
   Param,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { isNumber } from 'class-validator';
 
 @Controller('users')
 export class UsersController {
@@ -27,16 +29,23 @@ export class UsersController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
+    if (!isNumber(+id)) throw new BadRequestException('Invalid id');
+
     return this.userService.findOne(+id);
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    if (!isNumber(+id)) throw new BadRequestException('Invalid id');
+    if (!updateUserDto) throw new BadRequestException('Invalid request body');
+
     return this.userService.update(+id, updateUserDto);
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!isNumber(+id)) throw new BadRequestException('Invalid id');
+
     return this.userService.remove(+id);
   }
 }
